week_04_reasoning/src/requirement.txt
langgraph 
langchain_core 
langchain_openai 
python-dotenv
langchain-community 
tavily-python